tools/autotag/requirements.txt
pygithub ~= 1.55
gitpython ~= 3.1.24
packaging
